# perf(object-hash): avoid using array to just concatenate the string